Split shared_validation into checks/, tools/, data/ subpackages
Groups the 19 flat files by the categories the README already documented
(wired-into-the-gate checks, standalone manual tools, static config/data),
now expressed as real physical structure instead of just prose. Builds on
the uv-based editable install (previous commit) so every caller's import
path can be updated safely, verified by the test suite rather than by
hand-tracing sys.path behavior.

- shared_validation/checks/ — bible_sot, lint, text_checks, greek_hebrew_gloss,
  lexicon_check, lexicon_family_check, lexicon_source (+ lexicon_data/),
  family_check, family_resolver, scripture_check, run_report
- shared_validation/tools/ — content_length_report, check_gitignore_pycache,
  paren_balance
- shared_validation/data/ — gloss_format.json, native_script_ranges.json,
  no_latin_languages.json, versification_exceptions.json
- shared_validation/report.py stays at the package root (every subpackage
  depends on it)

Updated every caller (discovery/encounters scripts, devocionales_scripts,
11 test files) to the new shared_validation.checks.X / .tools.X import
paths, and every in-package relative import and data-file path reference
to match the new nesting depth.

Verified: uv run pytest tests/ -> 260 passed, 7 subtests passed (matches
baseline exactly, both before and after this change).
