+++ encounters/encounters_scripts/validate_family.py
@@ -38,7 +38,7 @@ from pathlib import Path
 sys.path.insert(0, str(Path(__file__).parent.parent.parent))
 from encounters_schema_checks import check_required_fields
 
-from shared_validation.family_check import run
+from shared_validation.checks.family_check import run
 
 ENCOUNTERS_DIR = Path(__file__).parent.parent
 INDEX_PATH = ENCOUNTERS_DIR / "index.json"
