+++ shared_validation/checks/bible_sot.py
