+++ shared_validation/checks/lexicon_data/strongs_greek.json
