+++ shared_validation/data/native_script_ranges.json
