+++ shared_validation/data/versification_exceptions.json
