+++ tests/test_business_rules_encounters.py
@@ -28,7 +28,7 @@ sys.path.insert(0, str(REPO_ROOT / "encounters" / "encounters_scripts"))
 import validate_encounters as ve  # noqa: E402
 
 from shared_validation.report import Report  # noqa: E402
-from shared_validation.text_checks import is_cognate  # noqa: E402
+from shared_validation.checks.text_checks import is_cognate  # noqa: E402
 
 # ── CARD_REQUIRED_KEYS dispatch ─────────────────────────────────────────────
 #
