+++ tests/test_lexicon_source.py
@@ -19,7 +19,7 @@ from pathlib import Path
 REPO_ROOT = Path(__file__).parent.parent
 sys.path.insert(0, str(REPO_ROOT))
 
-from shared_validation.lexicon_source import StrongsLexiconSource  # noqa: E402
+from shared_validation.checks.lexicon_source import StrongsLexiconSource  # noqa: E402
 
 
 class TestCaseInsensitiveLemmaLookup(unittest.TestCase):
