+++ tests/test_scripture_check.py
@@ -22,7 +22,7 @@ sys.path.insert(0, str(REPO_ROOT / "devocionales_scripts"))
 
 from verse_resolver import VerseResolver  # noqa: E402
 
-from shared_validation.scripture_check import (  # noqa: E402
+from shared_validation.checks.scripture_check import (  # noqa: E402
     FUZZY_MATCH_THRESHOLD,
     ScriptureRef,
     _is_intentional_truncation,
